Add files via upload
This repository contains code and resources for an encryption and decryption project implemented in C++. The project aims to provide a simple and secure way to encrypt and decrypt data using two algorithms: matrix algorithm and tree algorithm. Feel free to use the ppt to understand the flow of project.